% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airquality.R
\docType{data}
\name{airquality}
\alias{airquality}
\title{New York Air Quality Measurements}
\format{
A data frame with 153 rows and 6 variables:
\describe{
\item{Ozone}{Mean ozone in parts per billion from 1300 to 1500 hours at Roosevelt Island (ppb)}
\item{Solar.R}{Solar radiation in Langleys in the frequency band 4000--7700 Angstroms from 0800 to 1200 hours at Central Park (lang)}
\item{Wind}{Average wind speed in miles per hour at 0700 and 1000 hours at LaGuardia Airport (mph)}
\item{Temp}{Maximum daily temperature in degrees Fahrenheit at La Guardia Airport (degrees F)}
\item{Month}{Month of measurement (Month, 1-12)}
\item{Day}{Day of measurement (Day, 1-31)}
}
}
\source{
https://www.rdocumentation.org/packages/datasets/versions/3.6.2/topics/airquality
}
\usage{
airquality
}
\description{
Daily air quality measurements in New York, May to September 1973
}
\references{
Chambers, J. M., Cleveland, W. S., Kleiner, B. and Tukey, P. A. (1983) Graphical Methods for Data Analysis. Belmont, CA: Wadsworth.
}
\keyword{datasets}
