% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midel.R
\name{dimp}
\alias{dimp}
\title{Multiple Imputation and Deletion}
\usage{
dimp(original_data, mice_obj, y)
}
\arguments{
\item{mice_obj}{<TODO>}

\item{y}{The desired outcome variable.}

\item{x}{The raw dataset before imputation.}
}
\value{
The imputed datasets with outcomes set back to NA, if they were initially missing.
}
\description{
This function performs multiple imputation (including the outcome variable) and deletes any imputed outcome values prior to analysis.
}
\examples{
<TODO>
}
\author{
Group A
}
