% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toydata.R
\docType{data}
\name{toydata}
\alias{toydata}
\title{Example data}
\format{
A data frame with 3720 rows and 4 variables:
\describe{
\item{LitID}{Litter ID}
\item{Dose}{Dose of ethylene glycol (EG) administered}
\item{FetalWt}{Fetal weight}
\item{FetalMal}{Indicator for fetal malformation}
}
}
\source{
https://content.sph.harvard.edu/fitzmaur/ala2e/
}
\usage{
toydata
}
\description{
Armed Conflict Data, need to update the below
}
\references{
Price, C.J., Kimmel, C.A., Tyl, R.W. and Marr, M.C. (1985).
The developmental toxicity of ethylene glycol in rats and mice.
Toxicological Applications in Pharmacology, 81, 113-127.
}
\keyword{datasets}
