% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimp.R
\name{dimp}
\alias{dimp}
\title{Multiple Imputation and Deletion}
\usage{
dimp(mice_obj, y)
}
\arguments{
\item{mice_obj}{A `mids` object created using the `mice()` function.}

\item{y}{The desired outcome variable.}
}
\value{
The imputed datasets with outcomes set back to NA, if they were initially missing.
}
\description{
This function performs multiple imputation (including the outcome variable) and deletes any imputed outcome values prior to analysis.
}
\examples{
# Load in the mice library
library(mice)

# Create a mids object
data <- airquality
mice_obj <- mice(data, m = 5, maxit = 5, seed = 123)

mice_obj_deleted <- dimp(mice_obj, "Ozone")
}
\author{
Mark Asuncion, Ting Lin, John Fei, Luke Bai, Jason Dang
}
